데이터 로케이션 마커 추가/조회 기능.

+++ src/main/java/com/example/datahubwebsite/Models/Mapper/LocationMapper.java
@@ -13,8 +13,8 @@ public class LocationMapper implements RowMapper<Location> {
     public Location mapRow(ResultSet rs, int rowNum) throws SQLException {
 //        ArrayList<Location> arrLocation = new ArrayList<Location>();
         Location location = new Location();
-        location.setLat(rs.getInt("lat"));
-        location.setLng(rs.getInt("lng"));
+        location.setLat(rs.getDouble("lat"));
+        location.setLng(rs.getDouble("lng"));
         location.setFieldname(rs.getString("fieldname"));
         location.setDetail(rs.getString("detail"));
